+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add an option to set `PanelSettings` on `ConsoleSettings` [[#121](https://github.com/DarkRewar/BaseTool/issues/121)]
 - Allows numpad enter as return keycode for `Console` [[#120](https://github.com/DarkRewar/BaseTool/issues/120)]
+- Add `GenericObjectPool` to get object pools in hierarchy [[#123](https://github.com/DarkRewar/BaseTool/issues/123)]
 
 ## Fixes